+++ CHANGES.md
@@ -7,6 +7,13 @@
   graceful per-module fallbacks.
   https://github.com/joblib/threadpoolctl/pull/219
 
+- Added the ability to check whether a limiting API affects just the current
+  thread or the whole process. Mainly aimed at debugging and diagnostics, and
+  somewhat unreliable, it is therefore enabled by default only for command-line
+  usage.
+  https://github.com/joblib/threadpoolctl/pull/213
+
+
 - Only warn about simultaneous `libomp` and `libiomp` usage on Linux, where the
   incompatibility is known to cause crashes.